requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
streamlit>=1.25.0
pillow>=9.0.0
numpy>=1.21.0
tensorflow>=2.12.0
requests>=2.25.0
gdown>=4.6.0
python-multipart>=0.0.5
